web3-modal: FS restructure
