+++ packages/web3-modal/src/components/modals/AccountModal/index.tsx
@@ -7,7 +7,7 @@ import { useTheme } from 'styled-components'
 import { UserOptionsCtrlState } from '../../../controllers/types'
 import { useConnectDisconnect, usePstlWeb3Modal, useUserConnectionInfo } from '../../../hooks'
 import { useConnectedChainAndWalletLogo } from '../../../hooks/misc/useLogos'
-import { useDeriveAppType } from '../../../providers/utils/connectors'
+import { useDeriveAppType } from '../../../utils/connectors'
 import { PstlModalTheme } from '../../../theme'
 import { ConnectorEnhanced } from '../../../types'
 import { BaseModalProps, ModalId } from '../common/types'
