+++ packages/web3-modal/src/controllers/types/state.ts
@@ -3,7 +3,7 @@ import { type useConnect } from 'wagmi'
 
 import { PstlWeb3ModalProps } from '../../providers'
 import { ReadonlyChains } from '../../providers/types'
-import { AppType } from '../../providers/utils/connectors'
+import { AppType } from '../../utils/connectors'
 import { ChainImages } from '../../types'
 import { ConnectorOverrides } from '../../types/connectors'
 import { AnyTransactionReceipt } from '../TransactionsCtrl/types'
